+++ CHANGELOG.md
@@ -23,6 +23,7 @@
   with path to project's odoo config file
 - Added `odoo-helper install py-tools` command to install extra tools like pylint, flake8, ...
 - Added `odoo-helper server ps` command
+- Added `odoo-helper addons uninstall` command
 - `odoo-helper scaffold` have new features and subcommands:
   - `odoo-helper scaffold repo` create repository. place it in repo dir
   - `odoo-helper scaffold addon` create new addon. place it in repo and automaticaly link.